apply sliver
