+++ lib/views/about/about_screen.dart
@@ -2,11 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import '../../models/data_state.dart';
-import '../../models/site.dart';
 import '../../providers/site_repository.dart';
 import '../../providers/account_repository.dart';
 import '../widgets/section_header.dart';
-import '../account/select_site_sheet.dart';
+import '../account/select_site_item.dart';
 import '../app_localizations.dart';
 import '../account/login_section.dart';
 import 'guidance_section.dart';
@@ -21,12 +20,13 @@ class AboutScreen extends HookConsumerWidget {
     final t = AppLocalizations.of(context)!;
     final isMember = ref.watch(siteAccountRepositoryProvider) is Success;
     final site = ref.watch(siteRepositoryProvider);
-    final bool isSite = site is Success<(Site, List<Site>)>;
+    final bool isSite = site is Success<SiteRecord>;
     int index = 0;
 
     return CustomScrollView(
       slivers: [
-        if (!isMember) SelectSiteSheet(index: index++),
+        if (!isMember)
+          SliverToBoxAdapter(child: SelectSiteSheet(index: index++)),
         if (isSite && !isMember)
           SectionHeader(
             title: t.login,
