+++ lib/views/home/users_section.dart
@@ -2,44 +2,42 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import '../../config.dart';
-import '../../providers/groups_repository.dart';
+import '../../models/group.dart';
+import '../../providers/users_repository.dart';
 import '../widgets/modal_item.dart';
 import '../widgets/section_header.dart';
 import '../app_localizations.dart';
-import 'users_section.dart';
+import 'groups_section.dart';
 
-class GroupsSection extends HookConsumerWidget {
-  final String? user;
-  const GroupsSection({super.key, this.user});
+class UsersSection extends HookConsumerWidget {
+  final Group? group;
+  const UsersSection({super.key, this.group});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final t = AppLocalizations.of(context)!;
-    final items = [
-      (title: t.allMembers, data: null),
-      ...(user == null
-              ? ref.watch(groupsRepositoryProvider)
-              : ref
-                  .watch(groupsRepositoryProvider)
-                  .where((group) => group.users.contains(user)))
-          .map((group) => (title: group.name, data: group)),
-    ];
+    final users = group == null
+        ? ref.watch(usersRepositoryProvider)
+        : ref
+            .watch(usersRepositoryProvider)
+            .where((user) => group!.users.contains(user.id))
+            .toList();
 
     return SliverFixedExtentList(
       itemExtent: listItemHeight,
       delegate: SliverChildBuilderDelegate(
-        childCount: items.length,
+        childCount: users.length,
         (BuildContext context, int index) => ModalItem(
           index: index,
-          title: items[index].title,
+          title: users[index].name,
           trailing: const Icon(Icons.more_horiz),
           child: CustomScrollView(
             slivers: [
               SectionHeader(
-                title: t.users,
+                title: t.groupMembership,
                 leading: Icons.people,
               ),
-              UsersSection(group: items[index].data),
+              GroupsSection(user: users[index].id),
             ],
           ),
         ),
